Update RPG_GearTab.cs to show list view by default for non-human pawns
The default behaviour for non-human pawns when list view is not enabled is to show an empty board. Players like me usually need to toggle the list view in this case to see the gears for mechanoids, etc.

This change is to simply use list view when the pawn is not human.